post integration update. fixed the multi-particle int conversion

+++ src/himmelblau/func_F.py
@@ -12,12 +12,12 @@
 import numpy as np
 
 def func_F(X, NO_OF_OUTS=1):
-    F = np.zeros((NO_OF_OUTS))
+    F = np.zeros((NO_OF_OUTS)).astype(np.float64) 
     noErrors = True
     try:
         x = X[0]
         y = X[1]   
-        F[0] = (x**2 + y - 11)**2 + (x + y**2 - 7)**2
+        F[0] = (x**2. + y - 11.)**2 + (x + y**2 - 7.)**2
     except:
         noErrors = False
